+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Allow `registerCustomMetric` directly
+- Allow running custom tasks
 
 ## 7.0.0 - 2019-08-07
 - [**BC**] Replace `deriveToWithLog` with `deriveToWithApplication`, to allow more actions then just a logging.